Adding inter-request delay to constant partitioning tests
These tests turned out to be quite hard to pass when system throughput it high.
Tolerating churn under maximum load is difficult.
This commit rate-limits clients for these tests and lowers the maximum latency for a request significantly.

The new tests are not as clean as they ought to be.
There's quite a bit of code duplication at the moment.
I'll fix it at some later date.