Simplify search test DSL
This is a rather large change to the testing infrastructure but should be, for the most part, semantics preserving. I've re-arranged quite a bit of BaseJUnitTest to make the process of writing tests (especially search tests) much simpler, and the tests themselves more understandable.

In particular, I've added a new setting to searches called "goals", similar to the existing "invariants" and "prunes". The benefit is two-fold. First, it allows tests which utilize searches for progress to *also* search for invariant violations from the very start, more quickly pointing out errors if they exist. Second, it makes the tests much more readable, which should help students digest what is supposed to be happening. Instead of figuring out that what should be happening is the negation of some "invariant" the search test is expecting to be violating.

This commit also includes some minor code cleanup, as well as a way for generics to allow the test settings classes' setter methods to return their own type (to allow call chaining), even in the methods defined in the base class.