Removing unnecessary comment for single-threaded BFS
After 1d53df it is true that multi-threaded BFS happens completely
in-order. Moreover, if a goal is found at depth X, no later
goal-matching state will be reported in the search results.

However, I think it's fine to leave this test single-threaded without
the comment since the single-threaded version is so fast.