Allow client-specific workloads in the visual debugger.
This change only adds the feature for labs 0 through 3. The input
after -d should be in one of the following forms:
- "<# servers> <# clients> <single workload>"
- "<# servers> <# clients> <client1 workload> ... <clientn workload>"
where each workload is a comma-separated list of commands.

Adding this functionality for Lab 4 will be done in a future change;
there the format of the workload is different from labs 1-3 since we
also allow the user to provide a sequence of ShardMaster commands.

Tests: manually ran the visual debugger with the updated code in my
solution repository.
- Checked that a single workload can be given to all clients
- Checked that distinct workloads can be given to each client
- Checked for an error if user provides incorrect # of workloads