Allow client-specific workloads in lab 4's visual debugger.
The tricky part is that since we also allow a custom configuration
workload, we can't simply say that the format is

`<# groups> <# servers/group> <# shardmasters> <# clients>
    <client 1 workload> (... <client n workload>) <config workload>`.

Consider the following two inputs.

- 1 1 1 2 "APPEND:foo:bar" "JOIN:1"
- 1 1 1 2 "APPEND:foo:bar" "APPEND:foo:baz"

Both have the same number of args; without trying to parse the string
into a command, it's hard to determine that the first case is a custom
config workload with a uniform client workload, while the second case
is the default config workload with specific workloads per client.

To distinguish between these, this change requests that the config workload
should be prefixed with "config=", e.g.:

- 1 1 1 2 "APPEND:foo:bar" config="JOIN:1"

Manually tested: ran lab 4 visual debugger and confirmed that (a) we
can specify new configs and (b) we can specify custom workloads for
nodes.  Specifically, ran the following four commands:
./run-tests.py -l 4 -d 2 1 1 2 "APPEND:foo:bar"
./run-tests.py -l 4 -d 2 1 1 2 "APPEND:foo:bar" config="JOIN:1,JOIN:2,MOVE:1:4"
./run-tests.py -l 4 -d 2 1 1 2 "APPEND:foo:bar" "APPEND:foo:baz"
./run-tests.py -l 4 -d 2 1 1 2 "APPEND:foo:bar" "APPEND:foo:baz" config="JOIN:1,JOIN:2,MOVE:1:4"