Open the visual debugger when we assert a goal is found.
As there's no invariant violating trace, we open up a trace that gets
to the starting state of the BFS that failed.

There's still a `goalMatchingState` method that needs to be updated to
open the visual debugger in those cases (e.g., when you can't find a
View in lab 2). That will be fixed in the next change.

Tested: ran the tests with my solution code, but with the reply to the
client removed. Confirmed that the visual debugger is opened in the
start state of the BFS, and that we show the in-flight messages and
timers so that we can try to get the replies to the client.