Open visual debugger when goal not found (#20)
One of the more challenging test failures to debug is when a search-based "liveness" test fails, due to a BFS failing to find a goal state. Typically these BFS searches do not start in the initial state, but rather start from some previously computed state, e.g., the result of a BFS to find a state where a view has been fully initialized. Thus, students cannot simply open up the visual debugger and try to drive the system towards the failed goal, because they do not even know what the start state is.

I'm not really sure what students are expected to do in this situation. What I do when they come to me for help is manually edit the test file to print a trace to the starting state. Then I painstakingly manually replay this trace in the visual debugger (god help me when message are implicitly duplicated by the trace), and then from there I try to drive the system toward the final goal state, and understand why it fails to reach it.

This commit automates the process, by opening the visual debugger in the failed search's start state.

Co-authored-by: Logan Gnanapragasam <gnanabit@cs.washington.edu>
Co-authored-by: Ellis Michael <emichael@cs.washington.edu>