Client specific workloads (#19)
Some tests in the test suite use "non-uniform" workloads, where different clients send different sequences of commands. (For example, this is common with append commands to test linearizability.) In dslabs master, the visual debugger is not easy to use on such workloads, because the command-line interface assumes/requires uniform client workloads, where each client sends the same sequence of commands.

## Labs 0 - 3

The input
after -d should be in one of the following forms:
- "<# servers> <# clients> <single workload>"
- "<# servers> <# clients> <client1 workload> ... <clientn workload>"
where each workload is a comma-separated list of commands.

Tests: manually ran the visual debugger with the updated code in my
solution repository.
- Checked that a single workload can be given to all clients
- Checked that distinct workloads can be given to each client
- Checked for an error if user provides incorrect # of workloads

## Lab 4

The format is
`<# groups> <# servers/group> <# shardmasters> <# clients>
    <client 1 workload> (... <client n workload>) <config workload>`.

Manually tested: ran lab 4 visual debugger and confirmed that (a) we
can specify new configs and (b) we can specify custom workloads for
nodes.  Specifically, ran the following four commands:
./run-tests.py -l 4 -d 2 1 1 2 "APPEND:foo:bar"
./run-tests.py -l 4 -d 2 1 1 2 "APPEND:foo:bar" "JOIN:1,JOIN:2,MOVE:1:4"
./run-tests.py -l 4 -d 2 1 1 2 "APPEND:foo:bar" "APPEND:foo:baz"
./run-tests.py -l 4 -d 2 1 1 2 "APPEND:foo:bar" "APPEND:foo:baz" "JOIN:1,JOIN:2,MOVE:1:4"

Co-authored-by: Logan Gnanapragasam <gnanabit@cs.washington.edu>