Add a (slightly simplified) state graph for pingpong's 1st BFS.
The simplifications are:
- send 2 commands rather than 10
- omit some details about Node superclass, ClientWorker.

Also note that the path to the image works for the actual handout to
be distributed to students, but the image will not display correctly
in the official DSLabs directory (https://github.com/emichael/dslabs)
because the directory structure is different.

Also, there are some exercises at the end to think more about how the
checker works.