Show prunes and goals in DebuggerWindow.
If SearchSettings are provided to the VizClient, then invariants,
prunes, and goals will be shown to the left of the graph.

Switching from Map<...> to List<Pair<...>> is not necessary, but
StatePredicate doesn't implement equals, so I don't think we lose
anything by doing this (whereas it may be slightly unintuitive to have
a Map where we don't use the key to look things up).