Prevent actions from being taken according to SearchSettings.
Rather than hiding the undeliverable events, we enable/disable the
button next to the message/timer. This means that students shouldn't
be confused about why the message is showing up, and we can put some
information on the tooltip to explain why a message or timer isn't
deliverable according to the settings (but at the moment, the tooltips
are unchanged).