Incorporate SearchSettings in the new debugger (#29)
As described in https://github.com/emichael/dslabs/pull/20#issuecomment-1039365182, search tests often have settings to prevent some traces from being explored (e.g., by prunes, by declaring links/nodes inactive, by not delivering timers for some nodes, etc.). This change enforces the SearchSettings specified in a search test to help students to debug the search-based liveness tests.

This change is done in 3 parts.
- The VizClient must be given the SearchSettings so that it can give them to the DebuggerWindow. This is done in https://github.com/emichael/dslabs/commit/58d6be91f26826b0adf791111e18cb727946e239.
- The DebuggerWindow must accept the SearchSettings and show the prunes and goals to the left hand side. This is done in https://github.com/emichael/dslabs/commit/333c63c6a1ebd73946c5a4b2d1aa919d25d167af.
- In a SingleNodePanel, actions which are prohibited according to SearchSettings must be impossible to take. Moreover, if the search reaches an end state (goal matched, invariant violated, or prune reached), all actions should be impossible to take. This is done in https://github.com/emichael/dslabs/commit/d451e7269675d62f63e5e96a26f80d3e3d2743f3.

NOTES:

- Commits https://github.com/emichael/dslabs/commit/58d6be91f26826b0adf791111e18cb727946e239 and https://github.com/emichael/dslabs/commit/333c63c6a1ebd73946c5a4b2d1aa919d25d167af remove unused constructors and replace data structures. The commit messages explain why these changes were made, but they're not necessary for incorporating SearchSettings.
- This change doesn't yet add tooltips to explain why the messages cannot be delivered. At the moment, I don't see how to use the `SearchSettings` to get a string explanation of why a message/timer can't be delivered. But we should eventually add tooltips to explain why some messages/timers aren't deliverable.
- This change doesn't take the dropped network into account.We only `dropPendingMessages` in two tests in lab 3, so I think it's lower priority compared to this.

TESTS: manually started the debugger from various search test failures and confirmed that the invariant pane and actions behave correctly. I aimed to check the following:

- Goal pane behaves correctly: in [ping-pong's search test](https://github.com/emichael/dslabs/blob/633785340430c00f008569cf6cf169fcbd37cbb2/labs/lab0-pingpong/tst/dslabs/pingpong/PingTest.java#L122), I set the search to send 2 commands with a maxDepth of 2; when the search failed, I checked that we could navigate to a state where all workloads finished, that the goal pane on the left shows that All clients' workloads are finished, and that no actions can be taken in the debugger. 
- Invariant violation pane behaves correctly: with the same modifications to the pingpong search, I modified the client to not check received pong values. I checked that we could navigate to a state with an invariant violation, that the invariant pane on the left shows that the client did not receive the expected results, and that no actions can be taken in the debugger. I also checked that from the original search test (no depth restriction), if we don't check pong values in the client, the debugger shows the invariant violated and prevents any actions from being taken in the state with an invariant violation.
- Prune pane behaves correctly: from a valid solution to lab 2, I modified the client to never send results to clients. Running [part 2 test 17](https://github.com/emichael/dslabs/blob/d57df4281411a723e074bc467197bf06b01230f5/labs/lab2-primarybackup/tst/dslabs/primarybackup/PrimaryBackupTest.java#L700), I checked that if we take a sequence that causes the ViewServer to send a view with viewNum=3, then the debugger shows that the prune was reached and prevents actions from being taken.
- Messages on inactive links are not deliverable: same test as above, server3 is marked inactive by the test, and no messages from server 3 are deliverable.
- Timers on requested nodes are undeliverable: I took a valid solution to lab 3 and modified [test 23](https://github.com/emichael/dslabs/blob/dd15351e5fdf92491ffc93b7d3bce617196c06f5/labs/lab3-paxos/tst/dslabs/paxos/PaxosTest.java#L1001) to have a maxDepth of 1. I confirmed that timers on servers 1,2,5 were not deliverable.


* Take all SearchSettings in VizClient, not only the failed invariant.

To incorporate SearchSettings into the viz, we need to save it in the
VizClient. Removing the constructor accepting a StatePredicate is not
strictly necessary, but it can be computed from the SearchSettings, so
there's not really a reason to keep it.

* Show prunes and goals in DebuggerWindow.

If SearchSettings are provided to the VizClient, then invariants,
prunes, and goals will be shown to the left of the graph.

Switching from Map<...> to List<Pair<...>> is not necessary, but
StatePredicate doesn't implement equals, so I don't think we lose
anything by doing this (whereas it may be slightly unintuitive to have
a Map where we don't use the key to look things up).

* Prevent actions from being taken according to SearchSettings.

Rather than hiding the undeliverable events, we enable/disable the
button next to the message/timer. This means that students shouldn't
be confused about why the message is showing up, and we can put some
information on the tooltip to explain why a message or timer isn't
deliverable according to the settings (but at the moment, the tooltips
are unchanged).

* Make requested UI changes.

- Add "(Ignored States)" to the prune panel label.
- Replace the prune lock icon with an eye, and desaturate the prune
  warning color.
- Remove tooltip text when necessary.

Manually tested: checked that prune appears with new text, icon, and
is slightly desaturated, and also checked that tooltip text disappears
when we go back in the history from an invariant-violating state.

* Remove code duplication for updateState and save msg JButton.

* Allow exploration past invariants/goals, and add tooltips.

Tested by running lab 0's search test with 2 commands and max depth 2,
with both a valid solution and the incorrect solution where the
PingClient doesn't check the Pong value, and confirmed that we could
explore past goals.

For tooltips, used lab 2 part 2 test 17 and lab 3 test 23 to check
that tooltip contents gave correct explanations for why messages and
timers were not deliverable.

* Add a toggle to ignore restrictions from SearchSettings.

Tested by running lab 2 part 2 test 17: selected the option to ignore
search requirements and confirmed that (a) messages from server 3
could be delivered. Turned the option off (i.e., enforcing search
requirements again), navigated to a prune and confirmed that no
messages/timers could be delivered. From this pruned state, turned the
option back on and confirmed that actions could be taken.

Co-authored-by: Logan Gnanapragasam <gnanabit@cs.washington.edu>
Co-authored-by: Ellis Michael <emichael@cs.washington.edu>