Enforce most recent BFS' SearchSettings in new viz.
Tested: ran lab 2 part 2 test 17 with a PBServer who doesn't send
replies to the client, as outlined in
https://github.com/emichael/dslabs/pull/29, and confirmed that the
debugger still enforces prunes and network settings. Also, modified
the ViewServer to never add a backup even if there was an idle server
available, and ensured that prunes and goals are shown when the test
fails on `initView`.