Giving tests unique numbers based on part. Fixes #28.
Previously, when tests where filtered out the numbers of other tests
were recalculated. This commit assigns unique test numbers based on the
@Part annotation (if present) and the test name itself.

This commit also adds the @Lab annotation and looks up tests suites and
viz configs based on their annotations by searching through the
classpath. This obviates the need for the uniquely-named classes in the
testsuites and vizconfigs packages.