Add the ability to save and replay traces generated by invariant violations (#31)
This commit is the first pass for a trace save and replay feature. This
allows students to build up a custom set of "test cases" from failed
search tests. Serialization/de-serialization is done on a best-effort
basis; if students change message definitions in particular, or if the
core testing files change (which I might do at any time) then the replay
no longer works. While search tests are fairly robust as-is, this will
allow very rapid replay of failures and potential post-hoc visualization
of failures.