Add timers to viz timer boxes incrementally
This makes it so the GUI doesn't have to constantly repaint timer queue
boxes on every step. It also allows expanded timer objects to remain
expanded instead of getting collapsed.

This commit also highlights new timers, mimicking the logic of messages.

Finally, this commit disables the delivery of duplicate timers (which
are never deliverable).